Avatar handling is moved to a separate class
